fix: audit findings — semaphore logging, single-entry resolve, filterHeaders dedup, stream_abort status
- Add insertRejectedLog for SemaphoreQueueFullError and SemaphoreTimeoutError
  to ensure concurrency rejections have audit trail
- Refactor StreamProxy.onEnd() to go through terminal() single-entry resolve
  via deferred mode, eliminating dual resolve path
- Extract filterHeaders + SKIP_DOWNSTREAM to types.ts, remove duplication
  between transport.ts and stream-proxy.ts
- Mark stream_abort as "completed" (not "failed") in extractTrackStatus,
  reflecting that upstream succeeded before client disconnect
